+++ History.md
@@ -2,6 +2,9 @@
 ==================
   * Bump to 0.5.1
   * Reworked the call to wkhtmltopdf to take advantage of new options in wkhtmltopdf v 0.10.rc2
+  * added default cover options and default toc options to configuration
+  * added not to POST\_INSTALL\_ to use the static wkhtmltopdf
+
 
 2010-12-27 
 ==================